feat: unificar almacenamiento en un único .gob + flag --dir para consultas - Reemplaza el formato per-file (manifest.gob + chunks/*.gob) por un único   vectors.gob por rama (o vectors-{branch}.gob), reduciendo I/O de disco   en indexación y búsqueda - Remueve StorageManifest, dirtyFiles, loadPerFile, loadChunkFile,   writeChunkFile, writeManifestFile, cleanupTempFiles - saveLocked() escribe todo atómicamente con temp + rename - load() soporta StorageData → []ChunkRecord → []ChunkRecordLegacy (float64) - SwitchBranch() usa pathPrefix + suffix + .gob - Nuevo flag CLI --dir para sobreescribir el directorio raíz en   --query, --grep, --generate y --list-files - Actualiza callers (main.go, handlers.go, tests) y AGENTS.md

+++ pkg/indexer/indexer_test.go
@@ -158,7 +158,7 @@ func TestGetStats_WithData(t *testing.T) {
 	chunks := []chunker.Chunk{
 		{ID: "c1", FilePath: "/a.go", RelPath: "a.go", Language: "go", Content: "package a", FileHash: "h1"},
 	}
-	st.UpsertChunks(chunks, map[string][]float64{"c1": {1, 0, 0, 0}})
+	st.UpsertChunks(chunks, map[string][]float32{"c1": {1, 0, 0, 0}})
 
 	stats := idx.GetStats()
 	if stats.TotalChunks != 1 {
