+++ pkg/storage/simd/dot_arm64.go
@@ -3,8 +3,8 @@
 package simd
 
 // dot is the scalar fallback for ARM64 (no NEON-based implementation yet).
-func dot(a, b []float64) float64 {
-	var sum float64
+func dot(a, b []float32) float32 {
+	var sum float32
 	for i := range a {
 		sum += a[i] * b[i]
 	}
